pyDKB/*/communication: move StreamBuilder into a separate file.
Submodules of `communication` (like `consumer`) might need to use the
Builder, and thus it must be defined independently.
